+++ laravel/documentation/changes.md
@@ -50,6 +50,7 @@
 - Simplified the `paths.php` file.
 - Only write file caches if number of minutes is greater than zero.
 - Added `$default` parameter to Bundle::option method.
+- Fixed bug present when using Eloquent models with Twig.
 
 <a name="upgrade-3.2"></a>
 ## Upgrading From 3.1