+++ laravel/documentation/changes.md
@@ -2,8 +2,13 @@
 
 ## Contents
 
+<<<<<<< HEAD:laravel/documentation/changes.md
 - [Laravel 3.2](#3.2)
 - [Upgrading From 3.1](#upgrade-3.2)
+=======
+- [Laravel 3.1.9](#3.1.9)
+- [Upgrading From 3.1.8](#upgrade-3.1.9)
+>>>>>>> master:changes.md
 - [Laravel 3.1.8](#3.1.8)
 - [Upgrading From 3.1.7](#upgrade-3.1.8)
 - [Laravel 3.1.7](#3.1.7)
@@ -59,6 +64,16 @@
 
 <a name="upgrade-3.2"></a>
 ## Upgrading From 3.1
+
+- Replace the **laravel** folder.
+
+<a name="3.1.9"></a>
+## Laravel 3.1.9
+
+- Fixes cookie session driver bug that caused infinite loop on some occasions.
+
+<a name="upgrade-3.1.9"></a>
+## Upgrading From 3.1.8
 
 - Replace the **laravel** folder.